requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
pillow==10.2.0
pydantic==2.5.0
streamlit==1.28.2
python-multipart==0.0.6
